Move query params type and helper functions to single file
